refactor: communications
split code into smaller files, easier to deal with and edit
