+++ src/modbus.rs
@@ -1,6 +1,8 @@
-mod communications;
+mod buffer_communications;
+mod multi_register_communications;
 mod offsets;
 mod protocol;
+mod register_communications;
 pub mod types;
 
 pub use protocol::REQUEST_RESPONSE_DELAY;
